Implement adaptive precision
Precision is now calculated based on the zoom level. This is a good way
of ensuring that the higher zoom levels show all the data in the correct
places whilst also enabling the lower zoom levels to group data and
avoid showing insane numbers of points. The precision also alters how
many points we show as the number of buckets produced by the aggregation
changes. This is important because choosing too higher a precision could
cause some data to be missed off the map completely because it produces
too many buckets. Essentially it's a very careful balance and I have
estimated some values that seem to work ok!