Use a better geo_filter in the UTF Grid return
Prior to this we were returning a 1m radius point as the geo filter if
there were multiple records at the same point, now we're actually using
the bounding box of the bucket the records are all in.

This required a minor refactor, resulting in any reference in the tiles
code to "points" changing to "buckets". Makes a bit more sense now.